+++ src/signConfigureBaker.c
@@ -16,17 +16,17 @@ const ux_flow_step_t *ux_sign_configure_baker_commission[8];
 UX_STEP_NOCB(
     ux_sign_configure_baker_capital_step,
     bnnn_paging,
-    {.title = "Capital", .text = (char *) global.signConfigureBaker.displayCapital});
+    {.title = "Amount to stake", .text = (char *) global.signConfigureBaker.capitalRestakeDelegation.displayCapital});
 
 UX_STEP_NOCB(
     ux_sign_configure_baker_restake_step,
-    bnnn_paging,
-    {.title = "Restake earnings", .text = (char *) global.signConfigureBaker.displayRestake});
+    bn,
+    {"Restake earnings", (char *) global.signConfigureBaker.capitalRestakeDelegation.displayRestake});
 
 UX_STEP_NOCB(
     ux_sign_configure_baker_open_status_step,
-    bnnn_paging,
-    {.title = "Pool status", .text = (char *) global.signConfigureBaker.displayOpenForDelegation});
+    bn,
+    {"Pool status", (char *) global.signConfigureBaker.capitalRestakeDelegation.displayOpenForDelegation});
 
 UX_STEP_NOCB(ux_sign_configure_baker_keys_step, nn, {"Update baker", "keys"});
 
@@ -34,34 +34,35 @@ UX_STEP_CB(
     ux_sign_configure_baker_url_cb_step,
     bnnn_paging,
     sendSuccessNoIdle(),
-    {.title = "URL", .text = (char *) global.signConfigureBaker.url});
+    {.title = "URL", .text = (char *) global.signConfigureBaker.url.urlDisplay});
 
 UX_STEP_NOCB(
     ux_sign_configure_baker_url_step,
     bnnn_paging,
-    {.title = "URL", .text = (char *) global.signConfigureBaker.url});
+    {.title = "URL", .text = (char *) global.signConfigureBaker.url.urlDisplay});
 
 UX_STEP_CB(ux_sign_configure_baker_continue, nn, sendSuccessNoIdle(), {"Continue", "with transaction"});
 
-UX_STEP_NOCB(
-    ux_sign_configure_baker_empty_url_step,
-    bn,
-    {"Empty URL", ""});
+UX_STEP_NOCB(ux_sign_configure_baker_empty_url_step, bn, {"Empty URL", ""});
 
 UX_STEP_NOCB(
     ux_sign_configure_baker_commission_transaction_fee_step,
-    bnnn_paging,
-    {.title = "Transaction fee", .text = (char *) global.signConfigureBaker.transactionFeeCommissionRate});
+    bn,
+    {"Transaction fee", (char *) global.signConfigureBaker.commissionRates.transactionFeeCommissionRate});
 
 UX_STEP_NOCB(
     ux_sign_configure_baker_commission_baking_reward_step,
-    bnnn_paging,
-    {.title = "Baking reward", .text = (char *) global.signConfigureBaker.bakingRewardCommissionRate});
+    bn,
+    {"Baking reward", (char *) global.signConfigureBaker.commissionRates.bakingRewardCommissionRate});
 
 UX_STEP_NOCB(
     ux_sign_configure_baker_commission_finalization_reward_step,
-    bnnn_paging,
-    {.title = "Finalization reward", .text = (char *) global.signConfigureBaker.finalizationRewardCommissionRate});
+    bn,
+    {"Finalization reward", (char *) global.signConfigureBaker.commissionRates.finalizationRewardCommissionRate});
+
+bool hasCommissionRate() {
+    return ctx->hasTransactionFeeCommission || ctx->hasBakingRewardCommission || ctx->hasFinalizationRewardCommission;
+}
 
 /**
  * Dynamically builds and initializes the capital, restake earnings, pool status and
@@ -69,7 +70,7 @@ UX_STEP_NOCB(
  * - Ensures that the UI starts with the shared review transaction screens.
  * - Only displays the parts of the transaction that are set in the transaction, and skips
  *   any optional fields that are not included.
- * - If there are either the URL or commission rates in the transaction, then it shows a continue screen
+ * - If either the URL or commission rates are in the transaction, then it shows a continue screen
  *   at the end.
  */
 void startConfigureBakerDisplay() {
@@ -95,14 +96,9 @@ void startConfigureBakerDisplay() {
         ux_sign_configure_baker_first[index++] = &ux_sign_configure_baker_keys_step;
     }
 
-    // TODO: Fix display for keys. We probably want to show them so that an update to the
-    // keys is shown and does not just look like an empty transaction. Or at least we have
-    // to display something stating that the keys are updated.
-
     // If there are additional steps, then show continue screen. If this is the last step,
     // then show signing screens.
-    if (ctx->hasMetadataUrl || ctx->hasTransactionFeeCommission || ctx->hasBakingRewardCommission ||
-        ctx->hasFinalizationRewardCommission) {
+    if (ctx->hasMetadataUrl || hasCommissionRate()) {
         ux_sign_configure_baker_first[index++] = &ux_sign_configure_baker_continue;
     } else {
         ux_sign_configure_baker_first[index++] = &ux_sign_flow_shared_sign;
@@ -137,8 +133,7 @@ void startConfigureBakerUrlDisplay(bool lastUrlPage) {
     if (!lastUrlPage) {
         ux_sign_configure_baker_url[index++] = &ux_sign_configure_baker_url_cb_step;
     } else {
-
-        if (ctx->urlLength == 0) {
+        if (ctx->url.urlLength == 0) {
             ux_sign_configure_baker_url[index++] = &ux_sign_configure_baker_empty_url_step;
         } else {
             ux_sign_configure_baker_url[index++] = &ux_sign_configure_baker_url_step;
@@ -146,8 +141,7 @@ void startConfigureBakerUrlDisplay(bool lastUrlPage) {
 
         // If there are additional steps show the continue screen, otherwise go
         // to signing screens.
-        if (ctx->hasTransactionFeeCommission || ctx->hasBakingRewardCommission ||
-            ctx->hasFinalizationRewardCommission) {
+        if (hasCommissionRate()) {
             ux_sign_configure_baker_url[index++] = &ux_sign_configure_baker_continue;
         } else {
             ux_sign_configure_baker_url[index++] = &ux_sign_flow_shared_sign;
@@ -202,28 +196,36 @@ void startConfigureBakerCommissionDisplay() {
 #define P1_COMMISSION_RATES 0x05
 
 void handleCommissionRates(uint8_t *cdata, uint8_t dataLength) {
-    uint8_t rateLength;
-
     if (ctx->hasTransactionFeeCommission) {
-        rateLength =
-            parseCommissionRate(cdata, ctx->transactionFeeCommissionRate, sizeof(ctx->transactionFeeCommissionRate));
-        cdata += rateLength;
-        dataLength -= rateLength;
+        uint32_t rate = U4BE(cdata, 0);
+        fractionToText(
+            rate,
+            ctx->commissionRates.transactionFeeCommissionRate,
+            sizeof(ctx->commissionRates.transactionFeeCommissionRate));
+        cx_hash((cx_hash_t *) &tx_state->hash, 0, cdata, 4, NULL, 0);
+        cdata += 4;
+        dataLength -= 4;
     }
 
     if (ctx->hasBakingRewardCommission) {
-        rateLength =
-            parseCommissionRate(cdata, ctx->bakingRewardCommissionRate, sizeof(ctx->bakingRewardCommissionRate));
-        cdata += rateLength;
-        dataLength -= rateLength;
+        uint32_t rate = U4BE(cdata, 0);
+        fractionToText(
+            rate,
+            ctx->commissionRates.bakingRewardCommissionRate,
+            sizeof(ctx->commissionRates.bakingRewardCommissionRate));
+        cx_hash((cx_hash_t *) &tx_state->hash, 0, cdata, 4, NULL, 0);
+        cdata += 4;
+        dataLength -= 4;
     }
 
     if (ctx->hasFinalizationRewardCommission) {
-        rateLength = parseCommissionRate(
-            cdata,
-            ctx->finalizationRewardCommissionRate,
-            sizeof(ctx->finalizationRewardCommissionRate));
-        dataLength -= rateLength;
+        uint32_t rate = U4BE(cdata, 0);
+        fractionToText(
+            rate,
+            ctx->commissionRates.finalizationRewardCommissionRate,
+            sizeof(ctx->commissionRates.finalizationRewardCommissionRate));
+        cx_hash((cx_hash_t *) &tx_state->hash, 0, cdata, 4, NULL, 0);
+        dataLength -= 4;
     }
 
     if (dataLength != 0) {
@@ -265,15 +267,24 @@ void handleSignConfigureBaker(
         ctx->hasBakingRewardCommission = (bitmap >> 6) & 1;
         ctx->hasFinalizationRewardCommission = (bitmap >> 7) & 1;
 
-        // TODO: Determine state based on the above booleans. If any in the first
-        // section, then expect that part and so on...
+        if (ctx->hasCapital || ctx->hasRestakeEarnings || ctx->hasOpenForDelegation || ctx->hasKeys) {
+            ctx->state = CONFIGURE_BAKER_FIRST;
+        } else if (ctx->hasMetadataUrl) {
+            ctx->state = CONFIGURE_BAKER_URL_LENGTH;
+        } else if (hasCommissionRate()) {
+            ctx->state = CONFIGURE_BAKER_COMMISSION_RATES;
+        }
+
         sendSuccessNoIdle();
-    } else if (P1_FIRST_BATCH == p1) {
+    } else if (P1_FIRST_BATCH == p1 && ctx->state == CONFIGURE_BAKER_FIRST) {
         int lengthCheck = dataLength;
 
         if (ctx->hasCapital) {
             uint64_t capitalAmount = U8BE(cdata, 0);
-            amountToGtuDisplay(ctx->displayCapital, sizeof(ctx->displayCapital), capitalAmount);
+            amountToGtuDisplay(
+                ctx->capitalRestakeDelegation.displayCapital,
+                sizeof(ctx->capitalRestakeDelegation.displayCapital),
+                capitalAmount);
             cx_hash((cx_hash_t *) &tx_state->hash, 0, cdata, 8, NULL, 0);
             cdata += 8;
             lengthCheck -= 8;
@@ -285,9 +296,9 @@ void handleSignConfigureBaker(
             cdata += 1;
             lengthCheck -= 1;
             if (restake == 0) {
-                memmove(ctx->displayRestake, "No", 3);
+                memmove(ctx->capitalRestakeDelegation.displayRestake, "No", 3);
             } else if (restake == 1) {
-                memmove(ctx->displayRestake, "Yes", 4);
+                memmove(ctx->capitalRestakeDelegation.displayRestake, "Yes", 4);
             } else {
                 THROW(ERROR_INVALID_TRANSACTION);
             }
@@ -300,11 +311,11 @@ void handleSignConfigureBaker(
             lengthCheck -= 1;
 
             if (openForDelegation == 0) {
-                memmove(ctx->displayOpenForDelegation, "Open for all", 13);
+                memmove(ctx->capitalRestakeDelegation.displayOpenForDelegation, "Open for all", 13);
             } else if (openForDelegation == 1) {
-                memmove(ctx->displayOpenForDelegation, "Closed for new", 15);
+                memmove(ctx->capitalRestakeDelegation.displayOpenForDelegation, "Closed for new", 15);
             } else if (openForDelegation == 2) {
-                memmove(ctx->displayOpenForDelegation, "Closed for all", 15);
+                memmove(ctx->capitalRestakeDelegation.displayOpenForDelegation, "Closed for all", 15);
             } else {
                 THROW(ERROR_INVALID_TRANSACTION);
             }
@@ -336,15 +347,25 @@ void handleSignConfigureBaker(
             cx_hash((cx_hash_t *) &tx_state->hash, 0, cdata, 64, NULL, 0);
 
             // We delay the display until we get the aggregation key.
+            ctx->state = CONFIGURE_BAKER_AGGREGATION_KEY;
             sendSuccessNoIdle();
         } else {
             if (lengthCheck != 0) {
                 THROW(ERROR_INVALID_TRANSACTION);
             }
+
+            if (ctx->hasMetadataUrl) {
+                ctx->state = CONFIGURE_BAKER_URL_LENGTH;
+            } else if (hasCommissionRate()) {
+                ctx->state = CONFIGURE_BAKER_COMMISSION_RATES;
+            } else {
+                ctx->state = CONFIGURE_BAKER_END;
+            }
+
             startConfigureBakerDisplay();
             *flags |= IO_ASYNCH_REPLY;
         }
-    } else if (P1_AGGREGATION_KEY == p1) {
+    } else if (P1_AGGREGATION_KEY == p1 && ctx->state == CONFIGURE_BAKER_AGGREGATION_KEY) {
         if (!ctx->hasKeys || dataLength != 160) {
             THROW(ERROR_INVALID_TRANSACTION);
         }
@@ -356,49 +377,66 @@ void handleSignConfigureBaker(
         // Election Proof
         cx_hash((cx_hash_t *) &tx_state->hash, 0, cdata, 64, NULL, 0);
 
+        if (ctx->hasMetadataUrl) {
+            ctx->state = CONFIGURE_BAKER_URL_LENGTH;
+        } else if (hasCommissionRate()) {
+            ctx->state = CONFIGURE_BAKER_COMMISSION_RATES;
+        } else {
+            ctx->state = CONFIGURE_BAKER_END;
+        }
+
         startConfigureBakerDisplay();
         *flags |= IO_ASYNCH_REPLY;
-    } else if (P1_URL_LENGTH == p1) {
+    } else if (P1_URL_LENGTH == p1 && ctx->state == CONFIGURE_BAKER_URL_LENGTH) {
         if (!ctx->hasMetadataUrl) {
             THROW(ERROR_INVALID_TRANSACTION);
         }
 
-        ctx->urlLength = U2BE(cdata, 0);
-        if (ctx->urlLength > 2048) {
+        ctx->url.urlLength = U2BE(cdata, 0);
+        if (ctx->url.urlLength > 2048) {
             THROW(ERROR_INVALID_TRANSACTION);
         }
 
         cx_hash((cx_hash_t *) &tx_state->hash, 0, cdata, 2, NULL, 0);
 
-        if (ctx->urlLength == 0) {
+        if (ctx->url.urlLength == 0) {
             // If the url has length zero, we don't wait for the url bytes.
+            if (hasCommissionRate()) {
+                ctx->state = CONFIGURE_BAKER_COMMISSION_RATES;
+            } else {
+                ctx->state = CONFIGURE_BAKER_END;
+            }
+
             startConfigureBakerUrlDisplay(true);
             *flags |= IO_ASYNCH_REPLY;
         } else {
+            ctx->state = CONFIGURE_BAKER_URL;
             sendSuccessNoIdle();
         }
-    } else if (P1_URL == p1) {
-        // NOTE: We don't have to check the bool here, as the state is checked and
-        // can only be correct if the bool was set! Nifty!
-
-        // We cannot display strings that are so long... So we actually have to display
-        // one at a time then...
-        if (ctx->urlLength > dataLength) {
+    } else if (P1_URL == p1 && ctx->state == CONFIGURE_BAKER_URL) {
+        if (ctx->url.urlLength > dataLength) {
             cx_hash((cx_hash_t *) &tx_state->hash, 0, cdata, dataLength, NULL, 0);
-            ctx->urlLength -= dataLength;
-            memmove(ctx->url, cdata, dataLength);
+            ctx->url.urlLength -= dataLength;
+            memmove(ctx->url.urlDisplay, cdata, dataLength);
             startConfigureBakerUrlDisplay(false);
             *flags |= IO_ASYNCH_REPLY;
-        } else if (ctx->urlLength == dataLength) {
-            memmove(ctx->url, cdata, ctx->urlLength);
-            memmove(ctx->url + ctx->urlLength, "\0", 1);
-            cx_hash((cx_hash_t *) &tx_state->hash, 0, cdata, ctx->urlLength, NULL, 0);
+        } else if (ctx->url.urlLength == dataLength) {
+            memmove(ctx->url.urlDisplay, cdata, ctx->url.urlLength);
+            memmove(ctx->url.urlDisplay + ctx->url.urlLength, "\0", 1);
+            cx_hash((cx_hash_t *) &tx_state->hash, 0, cdata, ctx->url.urlLength, NULL, 0);
+
+            if (hasCommissionRate()) {
+                ctx->state = CONFIGURE_BAKER_COMMISSION_RATES;
+            } else {
+                ctx->state = CONFIGURE_BAKER_END;
+            }
+
             startConfigureBakerUrlDisplay(true);
             *flags |= IO_ASYNCH_REPLY;
         } else {
             THROW(ERROR_INVALID_TRANSACTION);
         }
-    } else if (P1_COMMISSION_RATES == p1) {
+    } else if (P1_COMMISSION_RATES == p1 && ctx->state == CONFIGURE_BAKER_COMMISSION_RATES) {
         handleCommissionRates(cdata, dataLength);
         *flags |= IO_ASYNCH_REPLY;
     } else {