+++ src/signConfigureDelegation.h
@@ -2,11 +2,11 @@
 #define _CONCORDIUM_APP_CONFIGURE_DELEGATION_H_
 
 /**
- * Handles the signing flow for an 'Configure Delegation' transaction. It validates
+ * Handles the signing flow for a 'Configure Delegation' transaction. It validates
  * that the correct UpdateType is supplied and will fail otherwise.
  * @param cdata please see /doc/ins_configure_delegation.md for details
  */
-void handleSignConfigureDelegation(uint8_t *cdata, volatile unsigned int *flags);
+void handleSignConfigureDelegation(uint8_t *cdata, uint8_t dataLength, volatile unsigned int *flags);
 
 typedef struct {
     uint8_t displayCapital[26];