+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Added flow to verify account address.
 * Added support for register data transaction.
 * Support for the configure delegation transaction.
+* Support for the configure baker transaction.
 
 ## 2.0.3 - 2021-11-29